Dockerisation complète avec PostgreSQL
- Dockerfile (Python 3.12-slim + Gunicorn), docker-compose.yml (web + PostgreSQL 16)
- docker-entrypoint.sh : attente DB + lancement Gunicorn (1 worker, 4 threads)
- Compatibilité PostgreSQL : BOOLEAN defaults, string quotes, json_build_object,
  table names quoting, rollback sur erreurs migration, pool engine conditionnel
- GameConfig.value élargi de VARCHAR(200) à TEXT (JSON planning trop long)
- func.round cast en Numeric pour PostgreSQL
- Compte admin créé en prod (plus limité au dev)
- .env.example, .dockerignore, docs mises à jour

+++ services/finance_service.py
@@ -1,6 +1,6 @@
 from datetime import datetime, timedelta
 
-from sqlalchemy import func, or_, update
+from sqlalchemy import func, or_, update, Numeric
 
 from data import EMERGENCY_RELIEF_AMOUNT, EMERGENCY_RELIEF_HOURS, EMERGENCY_RELIEF_THRESHOLD
 from extensions import db
@@ -35,7 +35,7 @@ def adjust_user_balance(user_id, delta, minimum_balance=None,
     stmt = update(User).where(User.id == user_id)
     if minimum_balance is not None:
         stmt = stmt.where(User.balance >= minimum_balance)
-    stmt = stmt.values(balance=func.round(User.balance + delta, 2)).returning(User.balance)
+    stmt = stmt.values(balance=func.round((User.balance + delta).cast(Numeric), 2)).returning(User.balance)
 
     row = db.session.execute(stmt).first()
     if not row:
@@ -132,7 +132,7 @@ def maybe_grant_emergency_relief(user):
             or_(User.last_relief_at.is_(None), User.last_relief_at <= cooldown_limit),
         )
         .values(
-            balance=func.round(User.balance + EMERGENCY_RELIEF_AMOUNT, 2),
+            balance=func.round((User.balance + EMERGENCY_RELIEF_AMOUNT).cast(Numeric), 2),
             last_relief_at=now,
         )
         .returning(User.balance)
